requirements.txt
flask
opencv-python
mediapipe
scikit-learn
numpy
pandas
joblib
Pillow